Validate directory access before updating preferences
This should fix the reported issues were some devices allow compressed files to be selected as directories, but do not allow write operations. This is done by checking if the selected URI is actually a file and, as a fallback, check if the name of the directory contains the extension of commonly found compressed files.